feat!: create root level HelmConfig struct which applies to all HelmReleases (#356)
First step into resolving
- https://github.com/validator-labs/validatorctl/issues/100

In an effort to simplify the TUI prompts in validatorctl to not require
lots of re-prompts for the helm configuration, we first need to refactor
our helm related structs.

The proposed structure would be this:

- `ValidatorConfigSpec.HelmConfig` defines the helm registry
configuration that would be used to pull all plugin charts
- `ValidatorConfigSpec.Plugins` has been simplifies to only specify the
charts repository, name, version, and values.

Along side this change, I updated the `reviewable-ext` make target to
also update the CRDs in the validator helm chart