fix: set oci.ImageOptions.Ref to the correct value to ensure plugin charts are installed (#359)
## Description
In validatorctl, when configuring a custom private registry, pulling the
plugin charts fails with this error:
```
2024-07-31 18:34:22 ERROR Plugin failed to install
                      └ validator-plugin-network: failed to pull chart: failed to fetch image from registry: GET https://toolbox.palette-adv.spectrocloud.com/v2/ahmad/charts/validator-plugin-network/validator-plugin-network/manifests/0.0.21: NOT_FOUND: repository ahmad/charts/validator-plugin-network/validator-plugin-network not found
```

As can be seen above, there is an extra `/validator-plugin-network`
tacked on at the end.
This PR fixes that.

This PR is a pre-requisite to
https://github.com/validator-labs/validatorctl/pull/115

## Test Notes
Tested this by:
1. running `helm install validator chart/validator/ -f
chart/validator/values.yaml`
2. running `./bin/validator install` and configuring a plugin using the
default registry (quay)
3. running `./bin/validator install` and configuring a plugin using a
custom private registry (harbor)

Everything runs correctly with this change. Prior to this change, test 3
was failing.